Feature/UI

+++ model/course.go
@@ -3,7 +3,7 @@ package model
 import (
 	"fmt"
 
-	util "github.com/eulersexception/glabs-ui/util"
+	"github.com/eulersexception/glabs-ui/util"
 	DB "modernc.org/ql"
 )
 
@@ -13,7 +13,6 @@ type Course struct {
 }
 
 func NewCourse(path string) *Course {
-
 	if path == "" {
 		fmt.Println("Enter valid course path.")
 		return nil
@@ -29,8 +28,8 @@ func NewCourse(path string) *Course {
 }
 
 func (c *Course) setCourse() {
-	db := util.GetDB()
-	defer util.FlushAndClose(db)
+	db := GetDB()
+	defer FlushAndClose(db)
 
 	_, _, err := db.Run(DB.NewRWCtx(), `
 		BEGIN TRANSACTION;
@@ -44,8 +43,8 @@ func (c *Course) setCourse() {
 }
 
 func GetCourse(path string) *Course {
-	db := util.GetDB()
-	defer util.FlushAndClose(db)
+	db := GetDB()
+	defer FlushAndClose(db)
 
 	rss, _, e := db.Run(nil, `
 		SELECT * FROM Course WHERE CoursePath = $1;
@@ -58,9 +57,7 @@ func GetCourse(path string) *Course {
 	c := &Course{}
 
 	for _, rs := range rss {
-
 		if er := rs.Do(false, func(data []interface{}) (bool, error) {
-
 			if err := DB.Unmarshal(c, data); err != nil {
 				return false, err
 			}
@@ -74,17 +71,85 @@ func GetCourse(path string) *Course {
 	return c
 }
 
-func DeleteSemesterFromCourse(path string) {
-	db := util.GetDB()
-	defer util.FlushAndClose(db)
+func UpdateCourse(oldPath string, newPath string) {
+	db := GetDB()
+	defer FlushAndClose(db)
 
 	_, _, err := db.Run(DB.NewRWCtx(), `
+		BEGIN TRANSACTION;
+			UPDATE Course CoursePath = $1 WHERE CoursePath = $2;
+			UPDATE Semester CoursePath = $1 WHERE CoursePath = $2;
+		COMMIT;
+	`, newPath, oldPath)
+
+	if err != nil {
+		panic(err)
+	}
+}
+
+func DeleteCourse(path string) {
+	semesters := GetAllSemestersForCourse(path)
+
+	if len(semesters) > 0 {
+		util.WarningLogger.Printf("Trying to delete course on path %s with existing references to semesters. First update course path on related semesters or delete related semesters.\n", path)
+	} else {
+		db := GetDB()
+		defer FlushAndClose(db)
+
+		_, _, err := db.Run(DB.NewRWCtx(), `
 		BEGIN TRANSACTION;
 			DELETE FROM Course WHERE CoursePath = $1;
 		COMMIT;
 	`, path)
 
+		if err != nil {
+			panic(err)
+		}
+	}
+}
+
+func (c *Course) AddSemesterToCourse(path string) {
+	db := GetDB()
+	defer FlushAndClose(db)
+
+	_, _, err := db.Run(DB.NewRWCtx(), `
+		BEGIN TRANSACTION;
+			UPDATE Semester CoursePath = $1 WHERE SemesterPath = $2;
+		COMMIT;
+	`, c.Path, path)
+
 	if err != nil {
 		panic(err)
 	}
+}
+
+func GetAllCourses() []Course {
+	db := GetDB()
+	rss, _, e := db.Run(nil, `SELECT * FROM Course;`)
+
+	if e != nil {
+		panic(e)
+	}
+
+	courses := make([]Course, 0)
+
+	for _, rs := range rss {
+		c := &Course{}
+
+		if er := rs.Do(false, func(data []interface{}) (bool, error) {
+			if err := DB.Unmarshal(c, data); err != nil {
+				return false, err
+			}
+
+			courses = append(courses, *c)
+
+			return true, nil
+		}); er != nil {
+			panic(er)
+		}
+	}
+
+	FlushAndClose(db)
+
+	return courses
 }
